requirements.txt
Django
djangorestframework
requests
pip install djangorestframework-jwt